pyproject.toml
[project]
name = "business_registry_scraper"
version = "0.0.1"
description = ""
authors = [
    { name = "Fauzaan Gasim", email = "hello@fauzaanu.com" }
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "crawlee",
    "apify",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]

